with images

+++ src/ua/tankv2/action/ActionField.java
@@ -2,6 +2,7 @@ package ua.tankv2.action;
 
 import ua.tankv2.field.Blank;
 import ua.tankv2.field.Water;
+import ua.tankv2.field.Rock;
 import ua.tankv2.managment.*;
 import ua.tankv2.tanks.*;
 import ua.tankv2.field.BattleField;
@@ -32,14 +33,14 @@ public class ActionField extends JPanel {
         aggressor   = new BT7(battleField,battleField.getAggressorLocation().get("x"),
                               battleField.getAggressorLocation().get("y"),Direction.DOWN);
 
-        tiger  = new Tiger(battleField,battleField.getTigerLocation().get("x"),
-                                       battleField.getTigerLocation().get("y"),
-                                       Direction.DOWN, 2);
+//        tiger  = new Tiger(battleField,battleField.getTigerLocation().get("x"),
+//                                       battleField.getTigerLocation().get("y"),
+//                                       Direction.DOWN, 2);
         List<Tank> listOfTank = new ArrayList<Tank>() {
             {
                 add(defender);
                 add(aggressor);
-                add(tiger);
+                //add(tiger);
             }
         };
 
@@ -99,7 +100,7 @@ public class ActionField extends JPanel {
         defender.draw(g);
         aggressor.draw(g);
         bullet.draw(g);
-        tiger.draw(g);
+        //tiger.draw(g);
 
     }
 
@@ -200,7 +201,7 @@ public class ActionField extends JPanel {
         Destroyable obf;
         if (y >= 0 && y < battleField.getDimentionY() && x >= 0 && x < battleField.getDimentionX()) {
             obf = battleField.scanQuadrant(y,x);
-            if (!obf.isDestroyed() && !(obf instanceof Blank) && !(obf instanceof Water)) {
+            if (!obf.isDestroyed() && !(obf instanceof Blank) && !(obf instanceof Water) && !(obf instanceof Rock)) {
                 battleField.destroyObject(y, x);
                 return true;
             }
